requirements.txt
pillow
google.genai
playwright
openai